Add files via upload
TECHNICAL REPORT FOR COMPLIX MOVIE STREAMING PERFORMANCE ANALYSIS 

Introduction - Complix Performance Analysis

Complix performance analysis typically involves evaluating how well the movie streaming organization is performing in various aspects of its business. The aims and objectives of Complix performance analysis revolve around using data-driven insights to optimize the company's performance, enhance user experience, and drive growth.

By leveraging performance analytics, Complix can make informed decisions to sustain its growth, retain customers, and offer a compelling entertainment experience. 

Problems being addressed

Some of the core areas of interest is to assess subscription growth, pricing strategies, and regional performance to improve profitability. Also, to analyze user behavior and preferences, enabling personalized recommendations that keep viewers engaged.

Key Dataset and Methodology 

The key dataset includes the subscription plan, regional performance, revenue by region, movie genre and movie ratings. Summarizes historical data to identify trends and patterns using Microsoft Excel. Sum function pivot table and chart was also employed in Microsoft Excel for this analysis.

Story of Data

The data used for this analysis is downloaded from Kaggle.com. According to Kaggle.com the dataset is a private dataset. 

Data Structure
The data structure has the column representing different variables such as customer name, revenue, region, subscription plan and the likes, while the rows represent different transactions by customers.
The important features in this performance analysis include the revenue, movie genre, region, customer name and subscription plan. The revenue and region are very important in determining the income generated from the viewing location and this helps to determine the best performing region. The subscription plan best used also assist us to know the plan of focus in terms of easy and seamless subscription.  

After observing the data manually, I discovered that the data was consistent to an appreciable level. There was no missing data in a cell, and no missing rows that needed cleaning. I also checked if data was duplicated in the dataset but no Data duplicate was discovered.

Data Splitting and Preprocessing
Data cleaning
Having checked and manually observed the data set, I discovered that the dataset was consistent. I then converted the dataset to a standard excel table for accuracy and easy automation.

I now Split the data into category 1 which is independent values and category 2 dependent values for the reliability and effectiveness of that analysis.

Data transformation 

There was a data transformation with the dataset I worked with and that includes adding variables like revenue and gender to the dataset in order to achieve proper performance analytics. 

Industry type and stakeholders

The industry type is a movie steaming organization and the stakeholder is the management team of the organization.

Value to the industry

The major value of this analysis to the industry is to keep track of their customer base growth and encourage existing customers in order to generate much more income since Revenue is their core value.

Pre-Analysis

Project split
Category 1 Independent Values
Customer Name, Country, Gender, Subscription Type, Favorite Genre
Category 2 Dependent Values
Age, User ID, Revenue, Watch time

Potential Analysis

Analysis of highest favorite genre by revenue
Analysis of highest patronizing country by revenue
Analysis of lead patronizing gender by revenue 
Analysis of highest subscription plan 
Analysis of the highest customers by revenue
Analysis of highest favorite genre by country

Potential Insight 
The highest favorite genre by revenues will help us know the movies genre that brings in more revenue so that priority will be placed on such genre of movies. 
Highest patronizing country by revenues will help us know the country that needs attention in terms of discounts and preference.
Analysis of the highest subscription plan helps us know the highest revenue generation plan thus we will make sure that subscription plan is always available and easy to subscribe to.. 

In Analysis

In analysis observation

I observe that the male gender uses this service more than the female gender. The dominant country where this service thrives is Canada. From the analysis, the highest customers of the same age bracket is the United states of America. Viewing the analysis dashboard, it is clear that the favorite genre of movie is Comedy. I observed also that the top subscription plan is the Premium Plan 

In analysis insight

Based on the observation, more movies that encourage the male gender should be published. meanwhile, movies that will encourage patronage from the female counterpart should also be encouraged in order to gain them

Canada has more patronage than the other countries, so preference should be made on their favorite genre to encourage them, besides some form of discount should be introduced in that location. However, what encourages the customers in Canada should be discovered and replicated in the other countries for better patronage. 

Based on the analysis, favorite genre is comedy, which means more of this genre of movies needs to be encouraged and made available in excess. We can also check out the favorite genres of movies from other countries so as to encourage them by publishing more of the same.  
Data Visualization and Chart 
 


From the above chart the top country that patronizes the movie streaming organization is Canada 



Highest favorite movie genre by revenue is still canada as shown in the chart above.




The highest 10 customers that top the dataset by revenue is indicated in the chart above. However, the top customer in the chart is James Martinez.


 

Highest subscribed plan is the premium plan as indicated on the chart.



The male gender top the chart indicating that the make gender patronizes the movie streaming platform more than the female gender.



The United States tops the list of the same age bracket that access this streaming platform.


Netflix Performance Analysis Dashboard





Recommendations and Observations
I observe that the male gender uses this service more than the female gender. The dominant country where this service thrives is Canada. From the analysis, the highest customers of the same age bracket is the United states of America. Viewing the analysis dashboard, it is clear that the favorite genre of movie is Comedy. I observed also that the top subscription plan is the Premium Plan 

Based on the observation more movies that encourage the male gender should be published meanwhile, movies that will encourage patronage from the female counterpart should also be encouraged in order to gain them

Canada has more patronage than the other countries, so preference should be made on their favorite genre to encourage them, besides some form of discount should be introduced in that location. However, what encourages the customers in Canada should be discovered and replicated in the other countries for better patronage. 

Based on the analysis, favorite genre is comedy, which means more of this genre of movies needs to be encouraged and made available in excess. We can also check out the favorite genre of movies from other countries so as to encourage them by publishing more of the same.